requirements.txt
streamlit
streamlit-webrtc
openai-whisper
sounddevice
numpy
torch
av
pydub
streamlit-audiorec
matplotlib